rename product model
