+++ src/app/models/product.ts
